pyproject.toml
[project]
name = "google-drive-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.184.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "typer>=0.12.0",
    "questionary>=2.0.0",
    "rich>=13.0.0",
    "markdown>=3.5.0",
]

[project.scripts]
gdrive-sync = "gdrive_sync.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
